Authorization module

+++ protected/controllers/ColorCodeController.php
@@ -28,6 +28,11 @@ class ColorCodeController extends Controller
 	{
 		return array(
 			array('allow',  // allow all users to perform 'index' and 'view' actions
+					'actions'=>array('getColorAjax', 'getParamAjax'),
+					'users'=>array('@'),
+			),
+			/*
+			array('allow',  // allow all users to perform 'index' and 'view' actions
 				'actions'=>array('index','view', 'getColorAjax', 'getParamAjax'),
 				'users'=>array('*'),
 			),
@@ -42,6 +47,7 @@ class ColorCodeController extends Controller
 			array('deny',  // deny all users
 				'users'=>array('*'),
 			),
+			*/
 		);
 	}
 
@@ -51,9 +57,13 @@ class ColorCodeController extends Controller
 	 */
 	public function actionView($id)
 	{
-		$this->render('view',array(
-			'model'=>$this->loadModel($id),
-		));
+		if (Yii::app()->authManager->checkAccess('viewColorCode', Yii::app()->user->id)) {
+			$this->render('view',array(
+				'model'=>$this->loadModel($id),
+			));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
+		}
 	}
 
 	/**
@@ -62,25 +72,29 @@ class ColorCodeController extends Controller
 	 */
 	public function actionCreate()
 	{
-		$model=new ColorCode;
-		$model->length = " "; // Omitted feature
-		// Uncomment the following line if AJAX validation is needed
-		$this->performAjaxValidation($model);
-		
-		// Create the color code
-		if(isset($_POST['ColorCode']))
-		{
-			$model->attributes=$_POST['ColorCode'];
-			$increment = isset($_POST['box'])? true:false ;
-			$model= ColorCodeController::setModelProperties($model, $increment) ["model"];
+		if (Yii::app()->authManager->checkAccess('createColorCode', Yii::app()->user->id)) {
+			$model=new ColorCode;
+			$model->length = " "; // Omitted feature
+			// Uncomment the following line if AJAX validation is needed
+			$this->performAjaxValidation($model);
 			
-			if($model->save())
-				$this->redirect(array('view','id'=>$model->color_code));
+			// Create the color code
+			if(isset($_POST['ColorCode']))
+			{
+				$model->attributes=$_POST['ColorCode'];
+				$increment = isset($_POST['box'])? true:false ;
+				$model= ColorCodeController::setModelProperties($model, $increment) ["model"];
+				
+				if($model->save())
+					$this->redirect(array('view','id'=>$model->color_code));
+			}
+	
+			$this->render('create',array(
+				'model'=>$model,
+			));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
 		}
-
-		$this->render('create',array(
-			'model'=>$model,
-		));
 	}
 	
 	public function setModelProperties ($model, $increment, $last=true) {
@@ -124,23 +138,27 @@ class ColorCodeController extends Controller
 	 */
 	public function actionUpdate($id)
 	{
-		$model=$this->loadModel($id);
-
-		// Uncomment the following line if AJAX validation is needed
-		$this->performAjaxValidation($model);
-
-		if(isset($_POST['ColorCode']))
-		{
-			
-			$model->attributes=$_POST['ColorCode'];
-			
-			if($model->save())
-				$this->redirect(array('view','id'=>$model->color_code));
+		if (Yii::app()->authManager->checkAccess('updateColorCode', Yii::app()->user->id)) {
+			$model=$this->loadModel($id);
+	
+			// Uncomment the following line if AJAX validation is needed
+			$this->performAjaxValidation($model);
+	
+			if(isset($_POST['ColorCode']))
+			{
+				
+				$model->attributes=$_POST['ColorCode'];
+				
+				if($model->save())
+					$this->redirect(array('view','id'=>$model->color_code));
+			}
+	
+			$this->render('update',array(
+				'model'=>$model,
+			));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
 		}
-
-		$this->render('update',array(
-			'model'=>$model,
-		));
 	}
 	
 	public function actionGetColorAjax(){
@@ -200,11 +218,15 @@ class ColorCodeController extends Controller
 	 */
 	public function actionDelete($id)
 	{
-		$this->loadModel($id)->delete();
-
-		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-		if(!isset($_GET['ajax']))
-			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+		if (Yii::app()->authManager->checkAccess('deleteColorCode', Yii::app()->user->id)) {
+			$this->loadModel($id)->delete();
+	
+			// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
+			if(!isset($_GET['ajax']))
+				$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
+		}
 	}
 
 	/**
@@ -212,10 +234,14 @@ class ColorCodeController extends Controller
 	 */
 	public function actionIndex()
 	{
-		$dataProvider=new CActiveDataProvider('ColorCode');
-		$this->render('index',array(
-			'dataProvider'=>$dataProvider,
-		));
+		if (Yii::app()->authManager->checkAccess('viewColorCode', Yii::app()->user->id)) {
+			$dataProvider=new CActiveDataProvider('ColorCode');
+			$this->render('index',array(
+				'dataProvider'=>$dataProvider,
+			));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
+		}
 	}
 
 	/**
@@ -223,14 +249,18 @@ class ColorCodeController extends Controller
 	 */
 	public function actionAdmin()
 	{
-		$model=new ColorCode('search');
-		$model->unsetAttributes();  // clear any default values
-		if(isset($_GET['ColorCode']))
-			$model->attributes=$_GET['ColorCode'];
-
-		$this->render('admin',array(
-			'model'=>$model,
-		));
+		if (Yii::app()->authManager->checkAccess('adminColorCode', Yii::app()->user->id)) {
+			$model=new ColorCode('search');
+			$model->unsetAttributes();  // clear any default values
+			if(isset($_GET['ColorCode']))
+				$model->attributes=$_GET['ColorCode'];
+	
+			$this->render('admin',array(
+				'model'=>$model,
+			));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
+		}
 	}
 
 	/**
