+++ protected/controllers/ColorController.php
@@ -27,6 +27,7 @@ class ColorController extends Controller
 	public function accessRules()
 	{
 		return array(
+				/*
 			array('allow',  // allow all users to perform 'index' and 'view' actions
 				'actions'=>array('index','view'),
 				'users'=>array('*'),
@@ -42,6 +43,7 @@ class ColorController extends Controller
 			array('deny',  // deny all users
 				'users'=>array('*'),
 			),
+			*/
 		);
 	}
 
@@ -51,9 +53,13 @@ class ColorController extends Controller
 	 */
 	public function actionView($id)
 	{
-		$this->render('view',array(
-			'model'=>$this->loadModel($id),
-		));
+		if (Yii::app()->authManager->checkAccess('viewColorCode', Yii::app()->user->id)) {
+			$this->render('view',array(
+				'model'=>$this->loadModel($id),
+			));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
+		}
 	}
 
 	/**
@@ -62,37 +68,42 @@ class ColorController extends Controller
 	 */
 	public function actionCreate()
 	{
-		$model=new Color;
-
-		// Uncomment the following line if AJAX validation is needed
-		$this->performAjaxValidation($model);
-
-		if(isset($_POST['Color']))
-		{
-			$model->attributes=$_POST['Color'];
-			$uploadedFile=CUploadedFile::getInstance($model,'color_img');
-            $fileName = "{$uploadedFile}";  //file name
-            $model->color_img = $fileName;
-            
-            if($model->validate())
-            {
-            	$path = Yii::app()->params['colorUploadPath'].$fileName;
-            	if (file_exists ($path))
-            		Yii::app()->user->setFlash("color_create", "This image file '{$path}' exists, kindly rename the file and upload it again");
-            	elseif ($uploadedFile->saveAs($path)) {
-					$model->save();
-					$this->redirect(array('view','id'=>$model->color_id));
-            	}
-            	else 
-            		Yii::app()->user->setFlash("color_create", "Cannot save this image '{$path}'");
-            		
-				
+		
+		if (Yii::app()->authManager->checkAccess('createColorCode', Yii::app()->user->id)) {
+			$model=new Color;
+	
+			// Uncomment the following line if AJAX validation is needed
+			$this->performAjaxValidation($model);
+	
+			if(isset($_POST['Color']))
+			{
+				$model->attributes=$_POST['Color'];
+				$uploadedFile=CUploadedFile::getInstance($model,'color_img');
+	            $fileName = "{$uploadedFile}";  //file name
+	            $model->color_img = $fileName;
+	            
+	            if($model->validate())
+	            {
+	            	$path = Yii::app()->params['colorUploadPath'].$fileName;
+	            	if (file_exists ($path))
+	            		Yii::app()->user->setFlash("color_create", "This image file '{$path}' exists, kindly rename the file and upload it again");
+	            	elseif ($uploadedFile->saveAs($path)) {
+						$model->save();
+						$this->redirect(array('view','id'=>$model->color_id));
+	            	}
+	            	else 
+	            		Yii::app()->user->setFlash("color_create", "Cannot save this image '{$path}'");
+	            		
+					
+				}
 			}
+	
+			$this->render('create',array(
+				'model'=>$model,
+			));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
 		}
-
-		$this->render('create',array(
-			'model'=>$model,
-		));
 	}
 
 	/**
@@ -102,34 +113,38 @@ class ColorController extends Controller
 	 */
 	public function actionUpdate($id)
 	{
-		$model=$this->loadModel($id);
-
-		// Uncomment the following line if AJAX validation is needed
-		$this->performAjaxValidation($model);
-
-		if(isset($_POST['Color']))
-		{
-			$_POST['Color']['color_img'] = $model->color_img;
-			$model->attributes=$_POST['Color'];
-			$uploadedFile=CUploadedFile::getInstance($model,'color_img');
-			
-			if($model->validate())
-            {
-            	if (!empty($uploadedFile)) {
-            		$fileName = "{$uploadedFile}";  //file name
-            		$model->color_img = $fileName;
-            		$uploadedFile->saveAs(Yii::app()->params['colorUploadPath'].$model->color_img);  // image will uplode to rootDirectory/banner/
-            	
-            	}
-				$model->save();
-            	$this->redirect(array('view','id'=>$model->color_id));
+		if (Yii::app()->authManager->checkAccess('updateColorCode', Yii::app()->user->id)) {
+			$model=$this->loadModel($id);
+	
+			// Uncomment the following line if AJAX validation is needed
+			$this->performAjaxValidation($model);
+	
+			if(isset($_POST['Color']))
+			{
+				$_POST['Color']['color_img'] = $model->color_img;
+				$model->attributes=$_POST['Color'];
+				$uploadedFile=CUploadedFile::getInstance($model,'color_img');
 				
+				if($model->validate())
+	            {
+	            	if (!empty($uploadedFile)) {
+	            		$fileName = "{$uploadedFile}";  //file name
+	            		$model->color_img = $fileName;
+	            		$uploadedFile->saveAs(Yii::app()->params['colorUploadPath'].$model->color_img);  // image will uplode to rootDirectory/banner/
+	            	
+	            	}
+					$model->save();
+	            	$this->redirect(array('view','id'=>$model->color_id));
+					
+				}
 			}
+	
+			$this->render('update',array(
+				'model'=>$model,
+			));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
 		}
-
-		$this->render('update',array(
-			'model'=>$model,
-		));
 	}
 
 	/**
@@ -139,11 +154,15 @@ class ColorController extends Controller
 	 */
 	public function actionDelete($id)
 	{
-		$this->loadModel($id)->delete();
-
-		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-		if(!isset($_GET['ajax']))
-			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+		if (Yii::app()->authManager->checkAccess('deleteColorCode', Yii::app()->user->id)) {
+			$this->loadModel($id)->delete();
+	
+			// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
+			if(!isset($_GET['ajax']))
+				$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
+		}
 	}
 
 	/**
@@ -151,10 +170,14 @@ class ColorController extends Controller
 	 */
 	public function actionIndex()
 	{
-		$dataProvider=new CActiveDataProvider('Color');
-		$this->render('index',array(
-			'dataProvider'=>$dataProvider,
-		));
+		if (Yii::app()->authManager->checkAccess('viewColorCode', Yii::app()->user->id)) {
+			$dataProvider=new CActiveDataProvider('Color');
+			$this->render('index',array(
+				'dataProvider'=>$dataProvider,
+			));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
+		}
 	}
 
 	/**
@@ -162,14 +185,18 @@ class ColorController extends Controller
 	 */
 	public function actionAdmin()
 	{
-		$model=new Color('search');
-		$model->unsetAttributes();  // clear any default values
-		if(isset($_GET['Color']))
-			$model->attributes=$_GET['Color'];
-
-		$this->render('admin',array(
-			'model'=>$model,
-		));
+		if (Yii::app()->authManager->checkAccess('adminColorCode', Yii::app()->user->id)) {
+			$model=new Color('search');
+			$model->unsetAttributes();  // clear any default values
+			if(isset($_GET['Color']))
+				$model->attributes=$_GET['Color'];
+	
+			$this->render('admin',array(
+				'model'=>$model,
+			));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
+		}
 	}
 
 	/**
