+++ protected/controllers/DCSNameController.php
@@ -27,6 +27,7 @@ class DCSNameController extends Controller
 	public function accessRules()
 	{
 		return array(
+				/*
 			array('allow',  // allow all users to perform 'index' and 'view' actions
 				'actions'=>array('index','view'),
 				'users'=>array('*'),
@@ -42,6 +43,7 @@ class DCSNameController extends Controller
 			array('deny',  // deny all users
 				'users'=>array('*'),
 			),
+			*/
 		);
 	}
 
@@ -51,9 +53,13 @@ class DCSNameController extends Controller
 	 */
 	public function actionView($id)
 	{
-		$this->render('view',array(
-			'model'=>$this->loadModel($id),
-		));
+		if (Yii::app()->authManager->checkAccess('viewDCSName', Yii::app()->user->id)) {
+			$this->render('view',array(
+				'model'=>$this->loadModel($id),
+			));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
+		}
 	}
 
 	/**
@@ -62,21 +68,25 @@ class DCSNameController extends Controller
 	 */
 	public function actionCreate()
 	{
-		$model=new DCSName;
-
-		// Uncomment the following line if AJAX validation is needed
-		$this->performAjaxValidation($model);
-
-		if(isset($_POST['DCSName']))
-		{
-			$model->attributes=$_POST['DCSName'];
-			if($model->save())
-				$this->redirect(array('view','id'=>$model->dcs_id));
+		if (Yii::app()->authManager->checkAccess('createDCSName', Yii::app()->user->id)) {
+			$model=new DCSName;
+	
+			// Uncomment the following line if AJAX validation is needed
+			$this->performAjaxValidation($model);
+	
+			if(isset($_POST['DCSName']))
+			{
+				$model->attributes=$_POST['DCSName'];
+				if($model->save())
+					$this->redirect(array('view','id'=>$model->dcs_id));
+			}
+	
+			$this->render('create',array(
+				'model'=>$model,
+			));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
 		}
-
-		$this->render('create',array(
-			'model'=>$model,
-		));
 	}
 
 	/**
@@ -86,21 +96,25 @@ class DCSNameController extends Controller
 	 */
 	public function actionUpdate($id)
 	{
-		$model=$this->loadModel($id);
-
-		// Uncomment the following line if AJAX validation is needed
-		$this->performAjaxValidation($model);
-
-		if(isset($_POST['DCSName']))
-		{
-			$model->attributes=$_POST['DCSName'];
-			if($model->save())
-				$this->redirect(array('view','id'=>$model->dcs_id));
+		if (Yii::app()->authManager->checkAccess('updateDCSName', Yii::app()->user->id)) {
+			$model=$this->loadModel($id);
+	
+			// Uncomment the following line if AJAX validation is needed
+			$this->performAjaxValidation($model);
+	
+			if(isset($_POST['DCSName']))
+			{
+				$model->attributes=$_POST['DCSName'];
+				if($model->save())
+					$this->redirect(array('view','id'=>$model->dcs_id));
+			}
+	
+			$this->render('update',array(
+				'model'=>$model,
+			));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
 		}
-
-		$this->render('update',array(
-			'model'=>$model,
-		));
 	}
 	
 	
@@ -111,11 +125,15 @@ class DCSNameController extends Controller
 	 */
 	public function actionDelete($id)
 	{
-		$this->loadModel($id)->delete();
-
-		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-		if(!isset($_GET['ajax']))
-			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+		if (Yii::app()->authManager->checkAccess('deleteDCSName', Yii::app()->user->id)) {
+			$this->loadModel($id)->delete();
+	
+			// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
+			if(!isset($_GET['ajax']))
+				$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
+		}
 	}
 
 	/**
@@ -123,10 +141,14 @@ class DCSNameController extends Controller
 	 */
 	public function actionIndex()
 	{
-		$dataProvider=new CActiveDataProvider('DCSName');
-		$this->render('index',array(
-			'dataProvider'=>$dataProvider,
-		));
+		if (Yii::app()->authManager->checkAccess('viewDCSName', Yii::app()->user->id)) {
+			$dataProvider=new CActiveDataProvider('DCSName');
+			$this->render('index',array(
+				'dataProvider'=>$dataProvider,
+			));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
+		}
 	}
 
 	/**
@@ -134,14 +156,18 @@ class DCSNameController extends Controller
 	 */
 	public function actionAdmin()
 	{
-		$model=new DCSName('search');
-		$model->unsetAttributes();  // clear any default values
-		if(isset($_GET['DCSName']))
-			$model->attributes=$_GET['DCSName'];
-
-		$this->render('admin',array(
-			'model'=>$model,
-		));
+		if (Yii::app()->authManager->checkAccess('adminDCSName', Yii::app()->user->id)) {
+			$model=new DCSName('search');
+			$model->unsetAttributes();  // clear any default values
+			if(isset($_GET['DCSName']))
+				$model->attributes=$_GET['DCSName'];
+	
+			$this->render('admin',array(
+				'model'=>$model,
+			));
+		} else {
+			throw new CHttpException(403,'You are not authorized to perform this action.');
+		}
 	}
 
 	/**
